enable msg threating

+++ internal/config/config_test.go
@@ -3,6 +3,7 @@ package config
 import (
 	"io/ioutil"
 	"os"
+	"strconv"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -81,6 +82,46 @@ func TestIncorrectConfig(t *testing.T) {
 		assert.EqualError(t, err, "incorrect websocket server")
 		configFile.Close()
 
+		os.Remove(configFile.Name())
+	}
+
+}
+
+func makeConfigDataWithStorage(envMattermostToken string, mattermostHttp string, mattermostWebsocket string, teamName string,
+	jiraHttp string, enableMsgThreting bool, storageType string) string {
+	configData := makeConfigData(envMattermostToken, mattermostHttp, mattermostWebsocket, teamName, jiraHttp)
+	configData += "enable_msg_threating: " + strconv.FormatBool(enableMsgThreting) + "\n"
+	configData += "storage_type: " + storageType + "\n"
+	return configData
+}
+
+func TestStorage(t *testing.T) {
+	configFile, err := ioutil.TempFile("", "config-*.txt")
+	t.Setenv("token", "token")
+	fileName := configFile.Name()
+	if err == nil {
+		configData := makeConfigDataWithStorage("token", "http://localhost:0000", "ws://localhost:0000", "team", "http://localhost:0000", true, "in_memory")
+		configFile.WriteString(configData)
+		configFile.Close()
+		config, err := LoadConfig(configFile.Name())
+		assert.Equal(t, nil, err)
+		assert.Equal(t, IN_MEMORY, config.StorageType)
+
+		configFile, _ = os.OpenFile(fileName, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 755)
+		configData = makeConfigDataWithStorage("token", "http://localhost:0000", "ws://localhost:0000", "team", "http://localhost:0000", true, "postgres")
+		configFile.WriteString(configData)
+		configFile.Close()
+		config, err = LoadConfig(configFile.Name())
+		assert.Equal(t, nil, err)
+		assert.Equal(t, POSTGRES, config.StorageType)
+
+		configFile, _ = os.OpenFile(fileName, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 755)
+		configData = makeConfigDataWithStorage("token", "http://localhost:0000", "ws://localhost:0000", "team", "http://localhost:0000", true, "other")
+		configFile.WriteString(configData)
+		configFile.Close()
+		config, err = LoadConfig(configFile.Name())
+		assert.NotEqual(t, nil, err)
+		assert.Equal(t, "incorrect storage type", err.Error())
 		os.Remove(configFile.Name())
 	}
